#306 - Removing existing RecogitoLogo and RecogitoCircleLogo; Renaming Logo component to RecogitoLogo
