--- src/components/Logo/index.ts
@@ -1,1 +1,0 @@
-export * from './Logo';
