+++ src/components/RecogitoLogo/RecogitoLogo.tsx
@@ -1,29 +1,90 @@
-export interface RecogitoLogoProps {
-  height?: number;
+import React from 'react';
 
-  width?: number;
+interface Props {
+  color: string;
+
+  contrastColor: string;
+
+  height: number;
+
+  width: number;
 }
 
-/**
- * Icon from https://www.svgrepo.com/svg/347710/comment-discussion
- * MIT-Licensed
- */
-export const RecogitoLogo = (props: RecogitoLogoProps) => {
-  const height = props.height || 22;
-
-  const width = props.width || 22;
-
-  return (
-    <svg
-      fill='#000000'
-      width={`${width}px`}
-      height={`${height}px`}
-      viewBox="0 0 24 24">
-      <path 
-        fillRule="evenodd" 
-        d="M1.75 1A1.75 1.75 0 000 2.75v9.5C0 13.216.784 14 1.75 14H3v1.543a1.457 1.457 0 002.487 1.03L8.061 14h6.189A1.75 1.75 0 0016 12.25v-9.5A1.75 1.75 0 0014.25 1H1.75zM1.5 2.75a.25.25 0 01.25-.25h12.5a.25.25 0 01.25.25v9.5a.25.25 0 01-.25.25h-6.5a.75.75 0 00-.53.22L4.5 15.44v-2.19a.75.75 0 00-.75-.75h-2a.25.25 0 01-.25-.25v-9.5z" />
-      <path 
-        d="M22.5 8.75a.25.25 0 00-.25-.25h-3.5a.75.75 0 010-1.5h3.5c.966 0 1.75.784 1.75 1.75v9.5A1.75 1.75 0 0122.25 20H21v1.543a1.457 1.457 0 01-2.487 1.03L15.939 20H10.75A1.75 1.75 0 019 18.25v-1.465a.75.75 0 011.5 0v1.465c0 .138.112.25.25.25h5.5a.75.75 0 01.53.22l2.72 2.72v-2.19a.75.75 0 01.75-.75h2a.25.25 0 00.25-.25v-9.5z" />
-    </svg>
-  );
-};
+export const RecogitoLogo = (props: Props) => (
+  <svg
+    width={props.width}
+    height={props.height}
+    viewBox='0 0 644 121'
+    fill='none'
+    xmlns='http://www.w3.org/2000/svg'
+  >
+    <path
+      fillRule='evenodd'
+      clipRule='evenodd'
+      d='M81.6502 74.7853C77.9955 73.9793 74.6974 72.2776 72.1124 70.2176L69.4383 74.0688C75.7671 78.9948 86.731 81.3235 94.5752 78.6366C103.489 76.2184 108.302 65.829 103.756 57.6788C98.4972 47.7372 84.8591 46.9312 77.9955 39.1392C75.3214 35.9149 76.0345 30.2724 79.5109 27.9437C82.4524 25.436 89.94 24.809 93.8621 26.3316C97.7841 27.4959 100.637 30.1828 103.756 32.2428V0H0V95.0266H38.151V121L60.168 103.266L57.4047 99.8631L42.5187 111.865V90.638H4.36775V4.3886H99.2995V23.7343C91.7227 19.7039 79.2435 19.883 73.8061 27.4064C69.26 33.5862 72.3799 42.0052 78.7086 45.5877C84.77 49.7972 93.6838 52.2154 98.5864 58.037C101.974 61.9778 101.35 68.3368 97.3384 72.188C94.5752 75.3227 85.9288 76.0392 81.6502 74.6958V74.7853Z'
+      fill={props.color}
+    />
+    <path
+      fillRule='evenodd'
+      clipRule='evenodd'
+      d='M46.7287 54.1565L56.1002 71.3061L69.0419 95.0102H91.355L116.078 113.686V95.0102H135V12.4048H109.563V16.8044H130.715V90.6106H111.705V104.887L92.8723 90.6106H71.7195L51.5483 53.9769C53.2441 53.7973 55.0292 53.1688 56.8143 52.0913C58.9563 50.8343 60.8306 49.0385 62.3479 46.6142C63.9545 44.1899 64.6685 41.1371 64.6685 37.5456C64.6685 33.954 63.8652 30.9012 62.2587 28.4769C60.7414 25.9629 58.5993 24.0773 56.1002 22.8203C53.5119 21.5632 50.7451 20.9347 47.7105 20.9347H29.5029V79.6564H33.9656V54.1565H46.818H46.7287ZM46.4609 50.116H33.8763V25.2445H46.4609C50.5666 25.2445 53.7797 26.2322 56.2787 28.2974C58.6886 30.2727 59.9381 33.3255 59.9381 37.6354C59.9381 41.9452 58.6886 45.0878 56.2787 47.0632C53.8689 49.1283 50.5666 50.116 46.4609 50.116Z'
+      fill={props.contrastColor}
+    />
+    <path
+      d='M622.572 65.1184C622.572 66.7865 622.916 68.2956 623.605 69.646C624.306 70.9963 625.266 72.0686 626.484 72.8629C627.702 73.6439 629.112 74.0345 630.714 74.0345C632.316 74.0345 633.726 73.6439 634.943 72.8629C636.175 72.0686 637.134 70.9963 637.823 69.646C638.511 68.2956 638.855 66.7865 638.855 65.1184C638.855 63.4504 638.511 61.9478 637.823 60.6107C637.134 59.2604 636.175 58.1881 634.943 57.3938C633.726 56.5995 632.316 56.2023 630.714 56.2023C629.112 56.2023 627.702 56.5995 626.484 57.3938C625.266 58.1881 624.306 59.2604 623.605 60.6107C622.916 61.9478 622.572 63.4504 622.572 65.1184ZM617.409 65.1184C617.409 63.2121 617.74 61.4381 618.402 59.7966C619.077 58.1418 620.017 56.6988 621.222 55.4676C622.427 54.2364 623.836 53.2767 625.452 52.5883C627.08 51.8999 628.834 51.5557 630.714 51.5557C632.62 51.5557 634.381 51.8999 635.996 52.5883C637.611 53.2767 639.014 54.2364 640.206 55.4676C641.41 56.6988 642.344 58.1418 643.006 59.7966C643.668 61.4381 643.999 63.2121 643.999 65.1184C643.999 67.038 643.674 68.8252 643.026 70.48C642.377 72.1215 641.457 73.5645 640.265 74.8089C639.074 76.0401 637.671 76.9999 636.055 77.6883C634.44 78.3767 632.66 78.7209 630.714 78.7209C628.741 78.7209 626.941 78.3767 625.313 77.6883C623.697 76.9999 622.294 76.0401 621.103 74.8089C619.924 73.5645 619.011 72.1215 618.362 70.48C617.727 68.8252 617.409 67.038 617.409 65.1184Z'
+      fill={props.color}
+    />
+    <path
+      d='M605.104 52.0322H610.069V78.2443H605.104V52.0322Z'
+      fill={props.color}
+    />
+    <path
+      d='M578.104 52.0322H585.332C588.125 52.0322 590.428 52.5949 592.242 53.7201C594.069 54.8322 595.433 56.3744 596.333 58.3469C597.233 60.3195 597.683 62.5766 597.683 65.1184C597.683 67.6734 597.233 69.9372 596.333 71.9097C595.433 73.8822 594.069 75.4311 592.242 76.5564C590.428 77.6816 588.125 78.2443 585.332 78.2443H578.104V52.0322ZM583.028 56.4605V73.816H584.895C586.603 73.816 588.006 73.4454 589.105 72.704C590.217 71.9494 591.044 70.9168 591.587 69.6062C592.143 68.2956 592.421 66.8129 592.421 65.1581C592.434 63.4901 592.169 62.0007 591.627 60.6901C591.084 59.3795 590.25 58.3469 589.124 57.5924C588.012 56.8378 586.603 56.4605 584.895 56.4605H583.028Z'
+      fill={props.color}
+    />
+    <path
+      d='M565.281 52.0322H570.246V68.1963C570.246 70.3542 569.815 72.2208 568.955 73.7962C568.108 75.3715 566.91 76.5895 565.361 77.45C563.825 78.2972 562.011 78.7209 559.92 78.7209C557.815 78.7209 555.981 78.2972 554.419 77.45C552.87 76.5895 551.666 75.3715 550.805 73.7962C549.958 72.2208 549.534 70.3542 549.534 68.1963V52.0322H554.479V67.9382C554.479 69.262 554.704 70.3939 555.154 71.3338C555.617 72.2605 556.259 72.9688 557.08 73.4586C557.901 73.9352 558.847 74.1735 559.92 74.1735C560.979 74.1735 561.912 73.9352 562.72 73.4586C563.527 72.9688 564.156 72.2605 564.606 71.3338C565.056 70.3939 565.281 69.262 565.281 67.9382V52.0322Z'
+      fill={props.color}
+    />
+    <path
+      d='M531.368 56.4208H524.795V52.0322H542.885V56.4208H536.332V78.2443H531.368V56.4208Z'
+      fill={props.color}
+    />
+    <path
+      d='M510.166 51.5557C511.741 51.5557 513.112 51.7344 514.277 52.0918C515.455 52.436 516.388 52.8067 517.077 53.2038C517.778 53.601 518.202 53.8591 518.347 53.9783L515.925 58.0094C515.713 57.8373 515.349 57.599 514.833 57.2945C514.33 56.9768 513.721 56.6988 513.006 56.4605C512.291 56.209 511.516 56.0832 510.682 56.0832C509.478 56.0832 508.511 56.3149 507.783 56.7782C507.068 57.2283 506.711 57.8836 506.711 58.7441C506.711 59.3266 506.903 59.8495 507.287 60.3129C507.671 60.763 508.233 61.1866 508.975 61.5838C509.729 61.9809 510.649 62.3979 511.735 62.8348C512.688 63.2055 513.595 63.6357 514.455 64.1255C515.329 64.6154 516.104 65.1912 516.779 65.8531C517.467 66.5151 518.01 67.2829 518.407 68.1566C518.817 69.0304 519.023 70.0299 519.023 71.1551C519.023 72.4128 518.764 73.5116 518.248 74.4515C517.745 75.3914 517.043 76.1791 516.143 76.8145C515.256 77.45 514.237 77.9266 513.085 78.2443C511.933 78.562 510.722 78.7209 509.451 78.7209C507.73 78.7209 506.181 78.5223 504.805 78.1252C503.441 77.7148 502.336 77.2779 501.488 76.8145C500.641 76.3512 500.145 76.0533 499.999 75.921L502.541 71.731C502.7 71.8501 502.998 72.0421 503.434 72.3069C503.884 72.5716 504.427 72.843 505.063 73.121C505.711 73.399 506.413 73.6373 507.168 73.8359C507.935 74.0345 508.717 74.1338 509.511 74.1338C510.941 74.1338 512.019 73.8359 512.748 73.2402C513.476 72.6444 513.84 71.87 513.84 70.9168C513.84 70.202 513.608 69.5731 513.145 69.0304C512.681 68.4876 512.026 67.9911 511.179 67.541C510.345 67.0777 509.345 66.6077 508.18 66.1312C507.015 65.6413 505.93 65.0721 504.924 64.4234C503.931 63.7615 503.13 62.9539 502.521 62.0008C501.912 61.0344 501.607 59.8495 501.607 58.4463C501.607 57.0562 501.998 55.8449 502.779 54.8123C503.573 53.7797 504.619 52.9788 505.917 52.4095C507.227 51.8403 508.644 51.5557 510.166 51.5557Z'
+      fill={props.color}
+    />
+    <path
+      d='M460.378 78.9873C456.72 78.9873 453.466 78.1664 450.614 76.5246C447.791 74.8828 445.573 72.6649 443.961 69.871C442.348 67.0482 441.541 63.8654 441.541 60.3226C441.541 56.7798 442.348 53.5826 443.961 50.731C445.573 47.8795 447.791 45.6184 450.614 43.9478C453.466 42.2772 456.72 41.4419 460.378 41.4419C464.008 41.4419 467.219 42.2772 470.013 43.9478C472.836 45.6184 475.039 47.8795 476.624 50.731C478.208 53.5826 479 56.7798 479 60.3226C479 63.8654 478.208 67.0482 476.624 69.871C475.039 72.6649 472.836 74.8828 470.013 76.5246C467.219 78.1664 464.008 78.9873 460.378 78.9873ZM460.378 72.5497C462.683 72.5497 464.685 72.0168 466.384 70.9511C468.112 69.8854 469.452 68.4308 470.402 66.5874C471.353 64.7151 471.828 62.6125 471.828 60.2794C471.828 57.9463 471.353 55.8437 470.402 53.9714C469.452 52.0992 468.112 50.6158 466.384 49.5213C464.685 48.4267 462.683 47.8795 460.378 47.8795C458.045 47.8795 456.015 48.4267 454.287 49.5213C452.558 50.6158 451.205 52.0992 450.225 53.9714C449.275 55.8437 448.8 57.9463 448.8 60.2794C448.8 62.6125 449.275 64.7151 450.225 66.5874C451.205 68.4308 452.558 69.8854 454.287 70.9511C456.015 72.0168 458.045 72.5497 460.378 72.5497Z'
+      fill={props.contrastColor}
+    />
+    <path
+      d='M409.736 42.4791H416.692V27.9189H424.167V42.4791H433.283V48.9167H424.167V66.458C424.167 68.5319 424.527 70.0585 425.247 71.0378C425.996 72.0171 426.99 72.5068 428.228 72.5068C429.323 72.5068 430.172 72.3052 430.777 71.9019C431.382 71.4987 431.742 71.225 431.857 71.081L434.839 76.6113C434.666 76.7553 434.205 77.0289 433.456 77.4322C432.707 77.8354 431.742 78.1955 430.561 78.5123C429.38 78.8292 427.998 78.9876 426.414 78.9876C423.648 78.9876 421.33 78.1523 419.458 76.4817C417.614 74.7823 416.692 72.1035 416.692 68.4455V48.9167H409.736V42.4791Z'
+      fill={props.contrastColor}
+    />
+    <path
+      d='M391.158 77.9506V42.479H398.676V77.9506H391.158ZM395.004 30.3383C393.65 30.3383 392.483 29.8631 391.504 28.9126C390.553 27.9333 390.078 26.7667 390.078 25.4129C390.078 24.0592 390.553 22.8926 391.504 21.9133C392.483 20.934 393.65 20.4443 395.004 20.4443C395.896 20.4443 396.717 20.6748 397.466 21.1356C398.244 21.5965 398.863 22.2014 399.324 22.9502C399.785 23.6703 400.015 24.4912 400.015 25.4129C400.015 26.7667 399.526 27.9333 398.546 28.9126C397.567 29.8631 396.386 30.3383 395.004 30.3383Z'
+      fill={props.contrastColor}
+    />
+    <path
+      d='M358.439 96.9175C355.645 96.9175 353.11 96.6294 350.835 96.0534C348.588 95.4773 346.615 94.6564 344.916 93.5907C343.245 92.5537 341.906 91.3296 340.898 89.9182L345.823 85.2088C346.37 85.9577 347.105 86.7498 348.027 87.5851C348.977 88.4492 350.259 89.1837 351.872 89.7886C353.485 90.3935 355.587 90.6959 358.18 90.6959C360.714 90.6959 362.86 90.1486 364.617 89.0541C366.374 87.9884 367.714 86.5338 368.635 84.6904C369.557 82.8469 370.018 80.7731 370.018 78.4688V77.3023H377.363V78.9441C377.363 82.8613 376.513 86.1449 374.814 88.7949C373.143 91.4736 370.882 93.4898 368.031 94.8436C365.179 96.2262 361.982 96.9175 358.439 96.9175ZM370.018 77.9503V71.0375C369.788 71.7576 369.111 72.7513 367.987 74.0187C366.893 75.286 365.381 76.4382 363.451 77.4751C361.55 78.4832 359.303 78.9873 356.711 78.9873C353.427 78.9873 350.446 78.2096 347.767 76.6542C345.117 75.0988 343 72.9241 341.416 70.1302C339.861 67.3074 339.083 64.0094 339.083 60.2362C339.083 56.4629 339.861 53.1793 341.416 50.3854C343 47.5626 345.117 45.3736 347.767 43.8182C350.446 42.234 353.427 41.4419 356.711 41.4419C359.274 41.4419 361.492 41.9028 363.364 42.8245C365.265 43.7462 366.778 44.8119 367.901 46.0217C369.053 47.2314 369.73 48.2683 369.932 49.1324V42.4788H377.363V77.9503H370.018ZM346.514 60.2362C346.514 62.7997 347.076 65.0176 348.199 66.8898C349.323 68.762 350.777 70.2166 352.563 71.2535C354.378 72.2617 356.308 72.7657 358.353 72.7657C360.542 72.7657 362.5 72.2473 364.228 71.2103C365.957 70.1734 367.325 68.7188 368.333 66.8466C369.341 64.9456 369.845 62.7421 369.845 60.2362C369.845 57.7303 369.341 55.5412 368.333 53.669C367.325 51.768 365.957 50.299 364.228 49.262C362.5 48.1963 360.542 47.6635 358.353 47.6635C356.308 47.6635 354.378 48.1819 352.563 49.2188C350.777 50.2558 349.323 51.7103 348.199 53.5826C347.076 55.4548 346.514 57.6727 346.514 60.2362Z'
+      fill={props.contrastColor}
+    />
+    <path
+      d='M311.953 78.9873C308.295 78.9873 305.04 78.1664 302.188 76.5246C299.366 74.8828 297.148 72.6649 295.535 69.871C293.922 67.0482 293.115 63.8654 293.115 60.3226C293.115 56.7798 293.922 53.5826 295.535 50.731C297.148 47.8795 299.366 45.6184 302.188 43.9478C305.04 42.2772 308.295 41.4419 311.953 41.4419C315.582 41.4419 318.794 42.2772 321.588 43.9478C324.41 45.6184 326.614 47.8795 328.198 50.731C329.782 53.5826 330.574 56.7798 330.574 60.3226C330.574 63.8654 329.782 67.0482 328.198 69.871C326.614 72.6649 324.41 74.8828 321.588 76.5246C318.794 78.1664 315.582 78.9873 311.953 78.9873ZM311.953 72.5497C314.257 72.5497 316.259 72.0168 317.958 70.9511C319.686 69.8854 321.026 68.4308 321.976 66.5874C322.927 64.7151 323.402 62.6125 323.402 60.2794C323.402 57.9463 322.927 55.8437 321.976 53.9714C321.026 52.0992 319.686 50.6158 317.958 49.5213C316.259 48.4267 314.257 47.8795 311.953 47.8795C309.62 47.8795 307.589 48.4267 305.861 49.5213C304.133 50.6158 302.779 52.0992 301.799 53.9714C300.849 55.8437 300.374 57.9463 300.374 60.2794C300.374 62.6125 300.849 64.7151 301.799 66.5874C302.779 68.4308 304.133 69.8854 305.861 70.9511C307.589 72.0168 309.62 72.5497 311.953 72.5497Z'
+      fill={props.contrastColor}
+    />
+    <path
+      d='M273.802 72.7657C276.049 72.7657 277.863 72.4057 279.246 71.6856C280.629 70.9655 281.493 70.447 281.838 70.1302L285.511 75.5309C285.223 75.7901 284.531 76.2077 283.437 76.7838C282.371 77.3311 280.989 77.8351 279.289 78.296C277.619 78.7569 275.718 78.9873 273.586 78.9873C270.331 78.9873 267.307 78.2672 264.513 76.827C261.719 75.358 259.458 73.2266 257.73 70.4326C256.03 67.6387 255.181 64.2255 255.181 60.193C255.181 56.1317 256.03 52.7185 257.73 49.9533C259.458 47.1594 261.719 45.0423 264.513 43.6022C267.307 42.162 270.331 41.4419 273.586 41.4419C275.689 41.4419 277.561 41.7011 279.203 42.2196C280.845 42.7092 282.198 43.2565 283.264 43.8614C284.33 44.4663 285.036 44.9271 285.381 45.244L281.665 50.6446C281.435 50.4142 280.989 50.0685 280.326 49.6077C279.692 49.118 278.828 48.6716 277.734 48.2683C276.639 47.8651 275.329 47.6635 273.802 47.6635C271.959 47.6635 270.158 48.1675 268.401 49.1756C266.673 50.155 265.247 51.5807 264.124 53.453C263.001 55.3252 262.439 57.5719 262.439 60.193C262.439 62.8141 263.001 65.0752 264.124 66.9762C265.247 68.8484 266.673 70.2886 268.401 71.2967C270.158 72.2761 271.959 72.7657 273.802 72.7657Z'
+      fill={props.contrastColor}
+    />
+    <path
+      d='M218.017 61.7484C218.046 63.8798 218.507 65.8241 219.4 67.5811C220.293 69.3381 221.618 70.7351 223.375 71.772C225.161 72.8089 227.35 73.3274 229.942 73.3274C232.131 73.3274 233.989 73.0106 235.515 72.3769C237.071 71.7144 238.338 70.9367 239.318 70.0438C240.297 69.1221 241.003 68.3012 241.435 67.5811L245.107 72.1176C244.156 73.5002 242.99 74.71 241.607 75.7469C240.254 76.7838 238.597 77.5759 236.639 78.1232C234.68 78.6992 232.289 78.9873 229.467 78.9873C225.665 78.9873 222.367 78.1808 219.573 76.5678C216.808 74.926 214.676 72.6649 213.178 69.7846C211.681 66.8754 210.932 63.5486 210.932 59.8041C210.932 56.3477 211.637 53.2369 213.049 50.4718C214.46 47.6779 216.505 45.4744 219.184 43.8614C221.891 42.2484 225.146 41.4419 228.948 41.4419C232.433 41.4419 235.458 42.1764 238.021 43.6454C240.614 45.0855 242.616 47.1306 244.027 49.7805C245.467 52.4304 246.187 55.5988 246.187 59.2857C246.187 59.5161 246.173 59.9338 246.144 60.5386C246.115 61.1147 246.086 61.518 246.058 61.7484H218.017ZM239.015 56.0885C238.986 54.9075 238.641 53.6114 237.978 52.2C237.344 50.7886 236.308 49.5789 234.867 48.5708C233.427 47.5626 231.483 47.0586 229.035 47.0586C226.5 47.0586 224.455 47.5482 222.9 48.5276C221.373 49.5069 220.25 50.7022 219.529 52.1136C218.809 53.4962 218.392 54.8211 218.277 56.0885H239.015Z'
+      fill={props.contrastColor}
+    />
+    <path
+      d='M163 20.9199H182.918C186 20.9199 188.794 21.568 191.299 22.8642C193.834 24.1315 195.85 26.0037 197.348 28.4808C198.875 30.9291 199.638 33.9247 199.638 37.4675C199.638 40.9816 198.932 43.8907 197.521 46.195C196.138 48.4993 194.468 50.2419 192.509 51.4228C190.551 52.575 188.736 53.2087 187.065 53.3239L201.193 77.9509H192.812L179.591 53.8423H170.388V77.9509H163V20.9199ZM170.388 48.3553H180.325C183.724 48.3553 186.532 47.5056 188.75 45.8062C190.968 44.1068 192.077 41.3848 192.077 37.6404C192.077 33.8959 190.968 31.2028 188.75 29.561C186.561 27.9192 183.767 27.0983 180.369 27.0983H170.388V48.3553Z'
+      fill={props.contrastColor}
+    />
+  </svg>
+)
