+++ src/layouts/login/LoginLayout.astro
@@ -1,6 +1,6 @@
 ---
 import { BrandFooter } from '@components/Branding/BrandFooter';
-import { Logo } from '@components/Logo';
+import { RecogitoLogo } from '@components/RecogitoLogo';
 import { getTranslations } from '@i18n';
 import HeaderLayout from '@layouts/HeaderLayout.astro';
 import config from 'src/config.json';
